% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densitymorph.R
\name{densitymorph}
\alias{densitymorph}
\title{Quasi-distance DensityMorph}
\usage{
densitymorph(X1, X2, Xref, K)
}
\arguments{
\item{X1}{A matrix of points, with each row corresponding to a point's coordinates in a multi-dimensional real space.}

\item{X2}{A second matrix of points, to be compared to X1. The number of columns and rows must be equal to those of X1.}

\item{Xref}{A pre-calculated reference self-PDLR distribution}

\item{K}{The number of the first K nearest neighbours (NN) to average over, defaults to K = 1}
}
\value{
A pair of real numbers measuring the quasi-distances from X1 to X2, and from X2 to X1
}
\description{
Calculates the quasi-distance DensityMorph between two set of points of identical dimension and number
}
